refactor: move ConfigBar modals and PWA Prompt out of Chain
