--- .env.production
@@ -1,2 +1,0 @@
-# Hostname of backend storing user configurations
-VITE_CONFIG_HOST=https://api.rezervo.no
